JMS Extract send logic to JmsMessageProducer
- created a separate JmsMessageProducer class to keep the stage logic
  more focussed on coordinating elements within the stage.
- grouped the logic for creating the execution context to use in the
  producer stage as well as in the consumer stage.
